+++ src/Config.hs
@@ -206,6 +206,8 @@ sourceFile Control = "control"
 sourceFile Departed = "departed"
 sourceFile DnD = "dnd"
 sourceFile PearlJam = "pearl_jam"
+sourceFile Rajnikanth = "rajnikanth"
+sourceFile Show = "show"
 
 mapSource :: Text -> Name
 mapSource "pearl_jam" = 'PearlJam
@@ -361,6 +363,8 @@ mapSource "yoda" = 'Yoda
 mapSource "zelda" = 'Zelda
 mapSource "measurement" = 'Measurement
 mapSource "opera" = 'Opera
+mapSource "rajnikanth" = 'Rajnikanth
+mapSource "show" = 'Show
 mapSource item = error $ "mapSource: Invalid argument passed " <> (show item)
 
 guessSourceFile :: SourceData -> Text -> FilePath